requirements.txt
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.2.1
numpy==2.3.3
pandas==2.3.3
scipy==1.16.2
scikit-learn==1.7.2
implicit==0.8.8   # ALS library for collaborative filtering
